Continue to detour or mission waypoint after brake/rotate
On the previous implementation, autocontinue after brake would direct the
vehicle to the previously sent *mission* waypoint even if the last received
command was a *detour* waypoint. In this implementation, the vehicle
autocontinues to the last received waypoint being it a detour waypoint or a
mission waypoint.

Also on the previous implementation, the vehicle would not autocontinue after
rotating. We have now implemented it to allow the vehicle to continue to a
previously sent detour or mission waypoint.

Signed-off-by: Guilherme Campos Camargo <guilherme.campos.camargo@intel.com>