Set a minimum value for waypoint altitude
When Ardupilot receives a mission item with an altitude that is too close to
ground (less than 10 centimeters), it simply ignores the new altitude. It
correctly changes the horizontal position of the vehicle - according to the new
mission item - but keeps the vertical position unchanged.

We have overridden this behavior by setting a minimum altitude in our mission
items, what means that when a mission item is sent through
"send_mission_waypoint()" or "send_detour_waypoint()", its altitude is first
set as max(0.1, altitude), to allow the vehicle to move to the lowest safe
altitude.

Signed-off-by: Guilherme Campos Camargo <guilherme.campos.camargo@intel.com>