System id on constructor and check_socket()
Now the system id can be passed as a parameter to the constructor. If the
system id is not provided, it is calculated according to the port to which the
provided socket is connected.

Also, check_socket() has been implemented to check if the socket provided to
the constructor is valid.

Signed-off-by: Guilherme Campos Camargo <guilherme.campos.camargo@intel.com>