px4.hh: revert to orignal content
This file is a copy of px4_custom_mode.h, from PX4 Firmware.
This patch reverts to the original content, to make it easier to keep
it sync'ed with the original file.
Changes reverted by this patch were minor tab/space and precompiler
directives. Also, the original license header should have been maintained.

Signed-off-by: Anselmo L. S. Melo <anselmo.melo@intel.com>