build: Make out of source build mandatory
Building mavlink-vehicles in source produces an error due to a conflict
between tests binary and tests directory. The recommendation is to
build it out of source and to avoid future issues, this patch makes
out of source builds mandatory, i.e. in source builds will be aborted
right after cmake execution.

Signed-off-by: Anselmo L. S. Melo <anselmo.melo@intel.com>