Remove preloaded waypoint requirement
So far our code needed a preloaded mission waypoint to be set on the
drone before our connection, but that prevents the usage of
mav-vehicles to set the mission waypoint.

This path removes the preloaded waypoint requirement and prevents
mav-vehicles from requesting mission items from and empty mission list.
This fixes a index out of bounds error in px4.